requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
colorama==0.3.7
decorator==4.0.10
funcsigs==1.0.2
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
mock==2.0.0
nose==1.3.7
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.0
pickleshare==0.7.3
Pillow==3.3.1
prompt-toolkit==1.0.5
ptyprocess==0.5.1
Pygments==2.1.3
python-termstyle==0.1.10
simplegeneric==0.8.1
six==1.10.0
sniffer==0.3.5
traitlets==4.2.2
wcwidth==0.1.7
wheel==0.26.
